feat: Add discarded items tab to presenter view
Introduces a new "Discarded" tab in the presenter view to allow creators to track items that were previously in the inventory but are no longer shown. This enhances management of game state and item lifecycle.

Also updates the `InventoryObject` type to include an optional `wasEverInInventory` flag, enabling better tracking of an object's history.