feat: Add dual inventory layout
Introduces support for a dual-column inventory layout. This allows for more structured organization of inventory items, especially in larger games.

New settings for inventory layout and titles have been added to the game configuration and are now exposed in the presentation view and editor. Inventory items can now also be assigned to a specific inventory slot.