feat(editor): Default new objects to be visible
When creating a new inventory object, it is now set to be initially visible in the room by default. This simplifies the user experience by reducing the number of clicks required for common object configurations.

The editor UI for managing object visibility has also been slightly adjusted to accommodate this change.