feat: Support global puzzles
Introduce an `isGlobal` flag to puzzles, allowing them to be defined once and accessible across all rooms. This enhances reusability and simplifies the management of puzzles that are not tied to a specific location within the escape room.

The Presenter View now aggregates global unsolved puzzles with local unsolved puzzles for display. The Editor has been updated to include this new option when creating or editing puzzles.